backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
pydantic
python-dotenv
redis
pika
python-jose[cryptography]
requests
